requirements.txt
asgiref==3.9.1
boto3==1.40.34
botocore==1.40.34
certifi==2025.8.3
charset-normalizer==3.4.3
cloudinary==1.44.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.10
jmespath==1.0.1
pillow==11.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
